WIP - Phantom_Phoenix plus AX12 test
This is a WIP, I am still debugging using the USB2AX with my
enhancements to run my PhantomX on a Linux machine, probably starting
with Raspberry PI.

AX12 test hangs when it tries to read in a parameter of a non-existent
servo.  Not sure if USB2AX code or the DXL code on the RPI that is not
properly handling a timeout.