Edison - ILI9341 try our own SPI transfer func
Experiment with our own function to communicate between the User mode
and the kernel mode.  The MRAA spi_write_buf function does a malloc to
read in data, that we later have to free, So trying a spi_transfer_buf
function that we can pass both read and write buffers.  Note: On writes
we can pass in NULL, which also speeds things up as the kernel does not
have to copy the data back, if we are not going to use it anyway