Secondary thread force exit if it does not exit on own
There were cases where ctrl-c was caught, but the program did not exit.
I think it was due to the join to worker thread and the worker thread
did not exit.  Added more code to hopefully get worker thread to exit,
plus added timed join, that if after 2 seconds, it will abort...