Joystick object detects joystick exiting
If the joystick should turn off,  we detect that and close the file
handle and setup to be able to detect it being reconnected.

Also found when we connect during a program run, the axis values may be
bogus, so since we know that we will receive an initial message for each
button and axis, we ignore these.

Also release memory that we allocated.