Servo 1->19, detect and correct 1 servo missing
Got tired of on occasion having a servo renumber itself back to #1, so I
renumbered servo #1 to #19.

Then at startup it tries to detect all of the servos.  If only one servo
is missing and servo #1 exists, it sets #1 to the missing servo number.

Nice thing is if you have stock PhantomX, it will see #1 and set it to
my new #19...