Allow Teensy for Arbotix...
Changes to allow my Teensy Arbotix board to act like Arbotix Pro with my
code.  In particular with the teensy we can set the baud rate of 1000000
directly.  Note: we can likewise do this for the USB2AX.  However code
is still in place if this fails, we do this indirect by setting to
standard baud: 38400 and then using IOCTLS to setup a non-standard baud
rate.  Note: the IOCTLS do not work to the Teensy (the set also fails on
USB2AX)